+++ Assets/Scripts/Player/PlayerController.cs
@@ -201,8 +201,8 @@ namespace StarterAssets
 
         private void Start()
         {
-            Cursor.lockState = CursorLockMode.Locked;
-            Cursor.visible = false;
+            Cursor.lockState = CursorLockMode.None;
+            Cursor.visible = true;
             _cinemachineTargetYaw = CinemachineCameraTarget.transform.rotation.eulerAngles.y;
             if (CinemachineCameraTarget != null)
             {